% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{A list of 1000 randomly generated numbes}
\format{
mtcars: Motor Trend Car Road Tests
\describe{
\item{x}{integer}
...
}
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars}
}
\usage{
mtcars
}
\description{
A list of 1000 randomly generated numbes
}
\keyword{datasets}
