% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars.R
\docType{data}
\name{x}
\alias{x}
\title{A list of 1000 randomly generated numbes}
\format{
Mtcars dataset from R
\describe{
\item{mpg}{numeric, 	Miles/(US) gallon}
\item{cyl}{numeric, Number of cylinders}
...
}
}
\usage{
x
}
\description{
A list of 1000 randomly generated numbes
}
\keyword{datasets}
